Split off variable name options to their own section
