+++ config/stub_templates/python/stub_config.toml
@@ -1,10 +1,5 @@
 name = "python"
-variable_format = "snake_case"
 source_file_ext = "py"
-allow_uppercase_vars = false
-keywords = [
-  "for"
-]
 aliases = ["py"]
 
 [type_tokens]
@@ -14,3 +9,10 @@ Float = "float"
 Bool = " != \"0\""
 String = "UNUSED"
 Word = "UNUSED"
+
+[variable_name_options]
+casing = "snake_case"
+allow_uppercase_vars = false
+keywords = [
+  "for"
+]
