+++ config/stub_templates/rust/stub_config.toml
@@ -1,13 +1,5 @@
 name = "rust"
-variable_format = "snake_case"
 source_file_ext = "rs"
-allow_uppercase_vars = false
-keywords = [
-  "as", "break", "const", "continue", "crate", "else", "enum", "extern", "false", "fn", "for", "if", "impl", "in", "let",
-  "loop", "match", "mod", "move", "mut", "pub", "ref", "return", "self", "Self", "static", "struct", "super", "trait", "true",
-  "type", "unsafe", "use", "where", "while", "async", "await", "dyn", "abstract", "become", "box", "do", "final", "macro", "override",
-  "priv", "typeof", "unsized", "virtual", "yield", "try", "macro_rules", "union", "'static"
-]
 
 [type_tokens]
 Int = "i32"
@@ -15,3 +7,12 @@ Long = "i64"
 Float = "f64"
 Bool = "i32"
 
+[variable_name_options]
+casing = "snake_case"
+allow_uppercase_vars = false
+keywords = [
+  "as", "break", "const", "continue", "crate", "else", "enum", "extern", "false", "fn", "for", "if", "impl", "in", "let",
+  "loop", "match", "mod", "move", "mut", "pub", "ref", "return", "self", "Self", "static", "struct", "super", "trait", "true",
+  "type", "unsafe", "use", "where", "while", "async", "await", "dyn", "abstract", "become", "box", "do", "final", "macro", "override",
+  "priv", "typeof", "unsized", "virtual", "yield", "try", "macro_rules", "union", "'static"
+]
